refactor(Product): 인수테스트 로직을 acceptance 로 이동
