+++ src/test/java/kitchenpos/application/acceptance/BaseAcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.application.acceptance;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
