+++ src/test/java/kitchenpos/application/acceptance/ProductAcceptanceTest.java
@@ -1,9 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.application.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.BaseAcceptanceTest;
+import kitchenpos.application.acceptance.BaseAcceptanceTest;
 import kitchenpos.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
